+++ AirLib/include/controllers/simple_flight/firmware/interfaces/IBoardInputPins.hpp
@@ -6,7 +6,7 @@ namespace simple_flight {
 
 class IBoardInputPins {
 public:
-    virtual float readChannel(uint8_t index) const = 0; //output -1 to 1
+    virtual float readChannel(uint16_t index) const = 0; //output -1 to 1
 };
 
 }
